monitoring/requirements.txt
alibabacloud-credentials==0.0.6
alibabacloud-dyvmsapi20170525==1.0.2
alibabacloud-endpoint-util==0.0.2
alibabacloud-rpc-util==0.0.3
alibabacloud-tea==0.1.0
alibabacloud-tea-rpc==0.0.3
alibabacloud-tea-util==0.1.2
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-dybaseapi==1.0.5
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
Django==2.1.4
django-cors-headers==3.2.1
djangorestframework==3.9.0
dnspython==1.16.0
eventlet==0.26.1
gevent==20.6.2
greenlet==0.4.16
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
kombu==4.6.11
monotonic==1.5
mysqlclient==1.4.6
pycparser==2.20
PyMySQL==0.9.3
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
uWSGI==2.0.18